requirements.txt
flask==3.0.2
flask-smorest==0.44.0
python-dotenv==1.0.1
sqlalchemy==2.0.29
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
passlib==1.7.4